requirements.txt
anyio==4.0.0
certifi==2023.7.22
exceptiongroup==1.1.3
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
python-dotenv==1.0.0
python-telegram-bot==20.5
sniffio==1.3.0
